fp/backend/pyproject.toml
[tool.poetry]
name = "source-coder"
version = "0.1.0"
description = ""
authors = ["Mateo Liberatore <liberatoremate98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.1"
flask = "^3.1.2"
requests = "^2.32.5"
flask-login = "^0.6.3"
jsonify = "^0.5"
werkzeug = "^3.1.3"
google-genai = "^1.38.0"
flask-cors = "^6.0.1"
python-dotenv = "^1.1.1"
dotenv = "^0.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
